:seedling: Update delegate commands to typescript
