+++ src/commands/delegate/votes.ts
@@ -16,16 +16,19 @@
 import { flags as flagParser } from '@oclif/command';
 import BaseCommand from '../../base';
 import { getAPIClient } from '../../utils/api';
-import { query } from '../../utils/query';
 import { SORT_FIELDS } from '../../utils/constants';
+import { query } from '../../utils/query';
 
 const MAXIMUM_LIMIT = 100;
+const DEFAULT_LIMIT = 10;
+const DEFAULT_OFFSET = 0;
+const DEFAULT_SORT = 'balance:desc';
 
 const VOTES_SORT_FIELDS = SORT_FIELDS.filter(
 	field => !field.includes('publicKey'),
 );
 
-const processFlagInputs = (limitStr, offsetStr, sortStr) => {
+const processFlagInputs = (limitStr: string, offsetStr: string, sortStr: string) => {
 	const limit = parseInt(limitStr, 10);
 	const offset = parseInt(offsetStr, 10);
 	const sort = sortStr ? sortStr.trim() : undefined;
@@ -42,7 +45,7 @@ const processFlagInputs = (limitStr, offsetStr, sortStr) => {
 	) {
 		throw new Error('Offset must be an integer and greater than or equal to 0');
 	}
-	if (!VOTES_SORT_FIELDS.includes(sort)) {
+	if (sort !== undefined && !VOTES_SORT_FIELDS.includes(sort)) {
 		throw new Error(`Sort must be one of: ${VOTES_SORT_FIELDS.join(', ')}`);
 	}
 
@@ -54,24 +57,59 @@ const processFlagInputs = (limitStr, offsetStr, sortStr) => {
 };
 
 export default class VotesCommand extends BaseCommand {
-	async run() {
+	static args = [
+		{
+			name: 'addresses',
+			required: true,
+			description: 'Comma-separated address(es) to get information about.',
+		},
+	];
+
+	static description = `
+	Gets votes information for given account(s) from the blockchain.
+	`;
+
+	static examples = [
+		'delegate:votes 13133549779353512613L',
+		'delegate:votes 13133549779353512613L,16010222169256538112L',
+		'delegate:votes 13133549779353512613L,16010222169256538112L --limit 20 --offset 5 --sort balance:asc --pretty',
+	];
+
+	static flags = {
+		...BaseCommand.flags,
+		limit: flagParser.string({
+			description: 'Limit applied to results.',
+			default: '10',
+		}),
+		offset: flagParser.string({
+			description: 'Offset applied to results.',
+			default: '0',
+		}),
+		sort: flagParser.string({
+			description: 'Fields to sort results by.',
+			default: DEFAULT_SORT,
+		}),
+	};
+
+	async run(): Promise<void> {
 		const {
-			args: { addresses },
+			args: { addresses: addressesStr },
 			flags: { limit: limitStr, offset: offsetStr, sort: sortStr },
 		} = this.parse(VotesCommand);
 
+		const addresses: ReadonlyArray<string> = addressesStr.split(',').filter(Boolean);
 		const { limit, offset, sort } = processFlagInputs(
-			limitStr,
-			offsetStr,
-			sortStr,
+			limitStr as string,
+			offsetStr as string,
+			sortStr as string,
 		);
 
 		const req = addresses.map(address => ({
 			query: {
 				address,
-				limit: limit || 10,
-				offset: offset || 0,
-				sort: sort || 'balance:desc',
+				limit: limit || DEFAULT_LIMIT,
+				offset: offset || DEFAULT_OFFSET,
+				sort: sort || DEFAULT_SORT,
 			},
 			placeholder: {
 				address,
@@ -84,37 +122,3 @@ export default class VotesCommand extends BaseCommand {
 	}
 }
 
-VotesCommand.args = [
-	{
-		name: 'addresses',
-		required: true,
-		description: 'Comma-separated address(es) to get information about.',
-		parse: input => input.split(',').filter(Boolean),
-	},
-];
-
-VotesCommand.flags = {
-	...BaseCommand.flags,
-	limit: flagParser.string({
-		description: 'Limit applied to results.',
-		default: '10',
-	}),
-	offset: flagParser.string({
-		description: 'Offset applied to results.',
-		default: '0',
-	}),
-	sort: flagParser.string({
-		description: 'Fields to sort results by.',
-		default: 'balance:desc',
-	}),
-};
-
-VotesCommand.description = `
-Gets votes information for given account(s) from the blockchain.
-`;
-
-VotesCommand.examples = [
-	'delegate:votes 13133549779353512613L',
-	'delegate:votes 13133549779353512613L,16010222169256538112L',
-	'delegate:votes 13133549779353512613L,16010222169256538112L --limit 20 --offset 5 --sort balance:asc --pretty',
-];
