+++ test/commands/delegate/get.test.ts
@@ -15,8 +15,8 @@
  */
 import { expect, test } from '@oclif/test';
 import * as config from '../../../src/utils/config';
-import * as print from '../../../src/utils/print';
-import * as api from '../../../src/utils/api';
+import * as printUtils from '../../../src/utils/print';
+import * as apiUtils from '../../../src/utils/api';
 import * as queryHandler from '../../../src/utils/query';
 
 describe('delegate:get', () => {
@@ -29,9 +29,9 @@ describe('delegate:get', () => {
 	const apiClientStub = sandbox.stub();
 	const setupTest = () =>
 		test
-			.stub(print, 'default', sandbox.stub().returns(printMethodStub))
+			.stub(printUtils, 'print', sandbox.stub().returns(printMethodStub))
 			.stub(config, 'getConfig', sandbox.stub().returns({ api: apiConfig }))
-			.stub(api, 'default', sandbox.stub().returns(apiClientStub));
+			.stub(apiUtils, 'getAPIClient', sandbox.stub().returns(apiClientStub));
 
 	describe('delegate:get', () => {
 		setupTest()
@@ -57,7 +57,7 @@ describe('delegate:get', () => {
 			.stub(queryHandler, 'query', sandbox.stub().resolves(queryResult))
 			.command(['delegate:get', username])
 			.it('should get an delegate info and display as an object', () => {
-				expect(api.default).to.be.calledWithExactly(apiConfig);
+				expect(apiUtils.getAPIClient).to.be.calledWithExactly(apiConfig);
 				expect(queryHandler.query).to.be.calledWithExactly(
 					apiClientStub,
 					endpoint,
@@ -97,7 +97,7 @@ describe('delegate:get', () => {
 			.stub(queryHandler, 'query', sandbox.stub().resolves(queryResult))
 			.command(['delegate:get', usernames.join(',')])
 			.it('should get delegates info and display as an array', () => {
-				expect(api.default).to.be.calledWithExactly(apiConfig);
+				expect(apiUtils.getAPIClient).to.be.calledWithExactly(apiConfig);
 				expect(queryHandler.query).to.be.calledWithExactly(
 					apiClientStub,
 					endpoint,
@@ -134,7 +134,7 @@ describe('delegate:get', () => {
 			.it(
 				'should get delegates info only using non-empty args and display as an array',
 				() => {
-					expect(api.default).to.be.calledWithExactly(apiConfig);
+					expect(apiUtils.getAPIClient).to.be.calledWithExactly(apiConfig);
 					expect(queryHandler.query).to.be.calledWithExactly(
 						apiClientStub,
 						endpoint,
